backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==2.0.1
